requirements.txt
python-dotenv
requests
aiohttp
cloudscraper
python-telegram-bot>=20.0
psutil